move NakadiRuntimeException with the others
