+++ src/main/java/org/zalando/nakadi/repository/TopicRepositoryCreator.java
@@ -3,7 +3,7 @@ package org.zalando.nakadi.repository;
 import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.Storage;
 import org.zalando.nakadi.domain.Timeline;
-import org.zalando.nakadi.exceptions.NakadiRuntimeException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
 import org.zalando.nakadi.exceptions.runtime.TopicRepositoryException;
 
 import java.util.List;
