+++ src/main/java/org/zalando/nakadi/repository/zookeeper/ZkChildrenCache.java
@@ -7,7 +7,7 @@ import org.apache.zookeeper.KeeperException;
 import org.echocat.jomon.runtime.concurrent.RetryForSpecifiedCountStrategy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.zalando.nakadi.exceptions.NakadiRuntimeException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
 
 import static org.apache.curator.framework.recipes.cache.PathChildrenCache.StartMode.BUILD_INITIAL_CACHE;
 import static org.echocat.jomon.runtime.concurrent.Retryer.executeWithRetry;
