+++ src/main/java/org/zalando/nakadi/service/subscription/zk/ZkSubscriptionImpl.java
@@ -5,7 +5,7 @@ import org.apache.curator.framework.api.GetChildrenBuilder;
 import org.apache.curator.framework.api.GetDataBuilder;
 import org.apache.zookeeper.WatchedEvent;
 import org.apache.zookeeper.Watcher;
-import org.zalando.nakadi.exceptions.NakadiRuntimeException;
+import org.zalando.nakadi.exceptions.runtime.NakadiRuntimeException;
 
 import java.util.List;
 import java.util.function.Function;
